input: changes no longer reset policy changes
The output of the Input Editor no longer is an input of RegoEditor.
This has the effect that changing the input no longer re-creates the
RegoEditor, which would have reset its contents to the defaults.

Instead, we now have a setter on RegoEditor's return object to control
its input. Furthermore, that setter forces a lint run, which uses that
input.

While doing this, also getting rid of a failed promise in the SQLEditor.
There, we're using a dummy lint run to trigger an update to its outputs
without requring a button of some sort, with the default wait behaviour:
If nothing changes for 3/4 of a second, the output is produced.